requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.9.5
django-allauth==0.25.2
django-bootstrap-themes==3.3.6
django-celery==3.1.17
django-crispy-forms==1.6.0
django-datatables-view==1.13.0
django-extensions==1.6.3
django-grappelli==2.8.1
django-pipeline==1.6.8
django-taggit==0.18.1
kombu==3.0.35
Naked==0.1.31
oauthlib==1.1.0
psycopg2==2.6.1
pycrypto==2.6.1
python-openid==2.2.5
pytz==2016.4
PyYAML==3.11
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.1
shellescape==3.4.1
six==1.10.0
suds==0.4
Werkzeug==0.11.9
